agents/publicist/requirements.txt
Flask>=2.3.0
gunicorn>=21.2.0
google-cloud-storage>=2.16.0
